Create Coursework1.md
this is the code for assignment 1 of the course, created by me.

It achieves importing a huge dataset, which is selectively read for a specific date region to save on working memory load, using the SQLDF package. SQLDF package is installed on the local machine at the beginning.

The data is processed for date and time format with the final stage being producing exploratory graphs using the plot() function.